adding the recipes section
We're introducing a new section called "recipes" where we want to show more detailed examples of how our API works and can be used.

I've also updated bootstrap from 3.2.0 to 3.3.5 and adjusted the navigation a bit to hold the new entry and remove the bug where the navigation has a break.