update carrier specific field lengths and getting started (#93)
To make it easier for readers to start developing for/with shipcloud we've updated the `Getting started` section on our homepage to make it clear that it doesn't need more than 3 steps to get started.

We've also found some incorrect values for carrier specific field lengths. And since I was already working at these entries I've added values for TNT which were missing until now.