added OpenAPI specification (#232)
We're switching from JSON-Schema to OpenAPI.
To make it easier for developers to generate mocks, stubs and work with our API
definitions, there will be a single OAI spec file from now on. You can either
download it and include/use it your own toolchain or use SwaggerUI for browsing
it from within our developer portal.