Updates gems and fixes jekyll build
With the current configuration, the jekyll gem raised an error because
when building the site. The base.css file contained HTML instead of css.
This could be fixed by removing the default layout setting in the jekyll
config, so we need to set the layout in every page.
We also updated the gemset to ruby 3.1.3 and amended the Rakefile to
provide tasks for all project-relevant commands.